test(jcode-chat): split attachment test fixtures
